+++ ij2tpl.ts
@@ -21,13 +21,22 @@ const enum TokenMember {
 	ELSE_BLOCK
 }
 
+const enum TokenString {
+	IF =	'?',
+	NOT =	'!',
+	ELSE =	'*',
+	END =	'/',
+	RAW =	'#',
+	COMMENT = '-'
+}
+
 const enum TokenType {
-	IF = 0,	// '?'
-	NOT,	// '!'
-	ELSE,	// '*'
-	END,	// '/'
+	IF = 0,
+	NOT,
+	ELSE,
+	END,
 	TEXT,
-	RAW,	// '#'
+	RAW,
 	FORMAT,
 	COMMENT
 }
@@ -43,13 +52,14 @@ interface IMap {
 	[index: number]: any;
 }
 
+// See https://github.com/microsoft/TypeScript/issues/14682
 const TokenTypeMap: IMap = {
-	'?':	TokenType.IF,
-	'!':	TokenType.NOT,
-	'*':	TokenType.ELSE,
-	'/':	TokenType.END,
-	'#':	TokenType.RAW,
-	'-':	TokenType.COMMENT
+	[TokenString.IF]:	TokenType.IF,
+	[TokenString.NOT]:	TokenType.NOT,
+	[TokenString.ELSE]:	TokenType.ELSE,
+	[TokenString.END]:	TokenType.END,
+	[TokenString.RAW]:	TokenType.RAW,
+	[TokenString.COMMENT]:	TokenType.COMMENT
 };
 
 // NOTE: if we use `IndentedTestRe` with capture-group directly, the `<string>.replace` method
@@ -109,14 +119,11 @@ export function tokenize(source: string, prefix: string, suffix: string): IToken
 		type_ = value[0];
 
 		switch (type_) {
-		case '?':
-		case '!':
-		case '*':
-		case '/':
-		case '-': // comment
-			// FIXME: we don't want to save comments!
-
-			// Remove section(or comment)'s indentation if exists
+		case TokenString.IF:
+		case TokenString.NOT:
+		case TokenString.ELSE:
+		case TokenString.END:
+			// Remove section's indentation if exists
 			if (token[TokenMember.TYPE] === TokenType.TEXT) {
 				if (IndentedTestRe.test(token[TokenMember.VALUE]))
 					token[TokenMember.VALUE] = token[TokenMember.VALUE].replace(IndentedWhiteSpaceRe, '');
@@ -132,20 +139,34 @@ export function tokenize(source: string, prefix: string, suffix: string): IToken
 					i += 1; // LF
 					break;
 				case '\r':
-					// Safe way for access a character in a string
-					i += source.charAt(i + 1) === '\n' ?
-						2 // CRLF
+					// Have next character?
+					i += i + 1 < l ?
+						// Yes, next character is LF?
+						source[i + 1] === '\n' ?
+							2 // Yes, then newline is CRLF
+							:
+							1 // No, then newline is CR
 						:
-						1 // CR
+						1 // No, then newline is CR
 					;
 					break;
 				}
 			}
 		// eslint-disable-line no-fallthrough
-		case '#':
+		case TokenString.RAW:
 			value = value.slice(1).trim();
-			token = [(TokenTypeMap[type_] as TokenType), value], tokens.push(token);
+			token = [TokenTypeMap[type_], value], tokens.push(token);
 			break;
+		case TokenString.COMMENT:
+			// Remove comment's indentation if exists
+			if (token[TokenMember.TYPE] === TokenType.TEXT) {
+				if (IndentedTestRe.test(token[TokenMember.VALUE]))
+					token[TokenMember.VALUE] = token[TokenMember.VALUE].replace(IndentedWhiteSpaceRe, '');
+
+				if(!token[TokenMember.VALUE])
+					tokens.pop(); // Drop the empty text ''
+			}
+			break; // Difference with section, we keep comments newline here
 		default:
 			token = [TokenType.FORMAT, value], tokens.push(token);
 			break;
@@ -341,7 +362,6 @@ export class Renderer {
 				// Check if it is non-values(null and undefined)
 				if (value != null)
 					buffer += value;
-
 				break;
 			case TokenType.FORMAT:
 				value = context.resolve(token[TokenMember.VALUE]);
@@ -368,13 +388,12 @@ export class Renderer {
 	}
 }
 
-// See https://github.com/microsoft/TypeScript/issues/14682
 const TokenTypeReverseMap: IMap = {
-	[TokenType.IF]:	'?',
-	[TokenType.NOT]:	'!',
-	[TokenType.ELSE]:	'*',
-	[TokenType.END]:	'/',
-	[TokenType.RAW]:	'#'
+	[TokenType.IF]:	TokenString.IF,
+	[TokenType.NOT]:	TokenString.NOT,
+	[TokenType.ELSE]:	TokenString.ELSE,
+	[TokenType.END]:	TokenString.END,
+	[TokenType.RAW]:	TokenString.RAW
 };
 
 function buildTree(tokens: IToken[]): IToken[] {
@@ -429,10 +448,10 @@ function buildTree(tokens: IToken[]): IToken[] {
 			if (sections.length > 0)
 				// Is parent section has initialized else-block?
 				collector = ((section = (sections[sections.length - 1] as IToken), isArray(section[TokenMember.ELSE_BLOCK])) ?
-					// Yes, then parent block is else-block.
+					// Yes, then parent block is else-block
 					section[TokenMember.ELSE_BLOCK]
 					:
-					// No, then parent block is (if-)block.
+					// No, then parent block is (if-)block
 					section[TokenMember.BLOCK]) as IToken[]
 				;
 			else